+++ test/src/ontariomnrf/predictor/trillium2026/TrilliumMortalityTest.java
@@ -41,7 +41,7 @@ public class TrilliumMortalityTest {
 
 	static class Trillium2026TreeImpl implements Trillium2026Tree, Trillium2026MortalityPlot {
 
-		private final double growthStepLengthYr;
+		private final int growthStepLengthYr;
 		private final double meanTminJanuaryCelsius;
 		private final double meanTempJuneToAugustCelsius;
 		private final double totalPrecMarchToMayMm;
@@ -56,7 +56,7 @@ public class TrilliumMortalityTest {
 		private final boolean dummyHarvest;
 		private final boolean planted;
 		
-		Trillium2026TreeImpl(double growthStepLengthYr,
+		Trillium2026TreeImpl(int growthStepLengthYr,
 				double meanTminJanuaryCelsius,
 				double totalPrecMarchToMayMm,
 				double meanTempJuneToAugustCelsius,
@@ -94,7 +94,7 @@ public class TrilliumMortalityTest {
 		public int getMonteCarloRealizationId() {return mcReal;}
 
 		@Override
-		public double getGrowthStepLengthYr() {return growthStepLengthYr;}
+		public int getGrowthStepLengthYr() {return growthStepLengthYr;}
 
 		@Override
 		public double getMeanMinimumJanuaryTemperatureCelsius(REpiceaClimateVariableInformation resolution) {return meanTminJanuaryCelsius;}
@@ -155,7 +155,7 @@ public class TrilliumMortalityTest {
 			reader = new CSVReader(filename);
 			Object[] record;
 			while ((record = reader.nextRecord()) != null) {
-				double growthStepLengthYr = Double.parseDouble(record[reader.getHeader().getIndexOfThisField("dt")].toString());
+				int growthStepLengthYr = Integer.parseInt(record[reader.getHeader().getIndexOfThisField("dt")].toString());
 				double meanTminJanuaryCelsius = Double.parseDouble(record[reader.getHeader().getIndexOfThisField("MeanTminJanuary")].toString());
 				double totalPrecMarchToMayMm = Double.parseDouble(record[reader.getHeader().getIndexOfThisField("TotalPrecMarchToMay")].toString());
 				double meanTempJuneToAugustCelsius = Double.parseDouble(record[reader.getHeader().getIndexOfThisField("MeanTempJuneToAugust")].toString());
